chore: moved layout to src folder
