+++ frontend/src/pages/Queues/index.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Queues = () => {
+	return <div></div>;
+};
+
+export default Queues;
