+++ frontend/src/routes/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter, Switch } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
 
-import LoggedInLayout from "../components/_layout";
+import LoggedInLayout from "../layout";
 import Dashboard from "../pages/Dashboard/";
 import Tickets from "../pages/Tickets/";
 import Signup from "../pages/Signup/";
@@ -11,6 +11,7 @@ import Connections from "../pages/Connections/";
 import Settings from "../pages/Settings/";
 import Users from "../pages/Users";
 import Contacts from "../pages/Contacts/";
+import Queues from "../pages/Queues/";
 import { AuthProvider } from "../context/Auth/AuthContext";
 import { WhatsAppsProvider } from "../context/WhatsApp/WhatsAppsContext";
 import Route from "./Route";
@@ -40,6 +41,7 @@ const Routes = () => {
 							<Route exact path="/contacts" component={Contacts} isPrivate />
 							<Route exact path="/users" component={Users} isPrivate />
 							<Route exact path="/Settings" component={Settings} isPrivate />
+							<Route exact path="/Queues" component={Queues} isPrivate />
 						</LoggedInLayout>
 					</WhatsAppsProvider>
 				</Switch>
